Change how test fs11 procedure calls are counted
The verification step depends on the number of
calls to mmul.

However with -O0 on ARM, the form of the call is
not "bl mmul". It's "blr x10", where the address
of mmul is loaded in x10. Therefore it's hard to
count calls to a specific function. Instead of
search in the assembly file, cound in .ll file
instead for ARM and POWER (or any compiler with
LLVM backend).