Add LL_ARG_INDIRECT_BUFFERED for aarch64 ABI compliance
LL_ARG_INDIRECT does not copy.

LL_ARG_BYVAL adds the "byval" attribute in the generated
LLVM. "byval" is more than just a compilation directive, it
actually changes the data layout and performs a copy. On x86 and
power this is the mechanism chosen to pass by value but on ARM it
is not ABI compliant and causes crashes in ABI tests. On ARM we
need to gain control over how the copy is done and do it ourselves
in the caller's scope.