Rearrange loop metadata handling
Refactor loop metadata so that loop ids are constructed and cached with cons_loop_id_md().

There are four pieces of metadata which may be added to loops:

* !{!"llvm.loop.vectorize", 0}
* !{!"llvm.loop.vectorize", 1}
* !{!"llvm.loop.unroll.disable}
* !{!"llvm.loop.unroll.enable"}

These are arranged through the functions:

* cons_novectorize_metadata
* cons_no_depchk_metadata
* cons_nounroll_metadata
* cons_unroll_metadata

respectively.

Previously, each of these functions would construct their own loop id metadata
node and cache it. For correct propagation of line number metadata for loops, it is
necessary to be able to append additional information to these loop ids.

This meant that loops appearing later in the program would reuse the loop id
metadata of prior loops. Even though this was not strictly correct, it likely
had few negative consequences, until now, where distinct loop ids are required
for line number information.

The intent of the code was to avoid creating many copies of redundant nodes
such as `!{!"llvm.loop.vectorize", 0}`.

This patch brings things inline with the intent: each function now creates and
caches only the attribute (not the ID), and a new function cons_loop_id_md()
caches loop IDs for the appropriate length of time.

Note that this patch preserves a bug: !llvm.mem.parallel_loop_access takes a
loop ID as a parameter; memory instructions within a language-level loop are
meant to be marked with the one loop ID representing the language-level loop.
However, the behaviour is to construct a new loop ID for each block, resulting
in one language-level-loop being marked with multiple loop IDs. The effect is
that the vectorizer may not respect `!dir$ ivdep` for loops with more than one
basic block. Fixing this is out of scope for now, so cons_loop_id_md() is
called in the new patch exactly where new metadata would have been created
prior to this patch.

The change *does* include a minor functional change: multiple metadata may now
be specified for a single loop, where previously the last-one-assigned would
win.