Add debug line number information to loop metadata
LLVM 9's Loop simplification and canonicalization passes propagate line
number information in a different way than previously.

In order for !dbg line number information to survive these passes, the
!llvm.loop directive can contain two DILocation metadata nodes. The
first is used as the start line, and the second is used as the end line.

This patch adds two tests on the -emit-flang-llvm output.

* Test that the line number information is correct.
* Test that the line number and pragma metadata is correct when both are
  present.

The tests alternate between a variety of different pragmas to catch
issues with the caching of pragmas that flang does.