Fix a segfault with procedure dummy arguments
Fix a fault with procedure dummy arguments and interfaces.
Create a unique symbol for interfaces that are used in procedure
pointer declarations because the LLVM bridge uses the interface
as well as the argument ILI to generate arguments at a procedure
pointer call.  An interface symbol might turn into a procedure
symbol in the back-end.  Create a unique symbol so we can
mark it with a new field called IS_PROC_PTR_IFACE. This
field says that the symbol is used as an interface to a
procedure pointer. When the bridge sees this field set, it
knows to generate a closure pointer as the last argument at
the call-site.

This bug fix addresses Flang issue #577.