Change Flang licensing to Apache 2.0 with LLVM-Exceptions
With the changes to LLVM licensing to the Apache 2.0
license with LLVM exceptions, we are changing the
flang file copyright headers to use the new wording that
has been adopted by LLVM.

The header only contains the license information, and the
copyright and dates have been dropped.

See http://llvm.org/docs/DeveloperPolicy.html#new-llvm-project-license-framework
for a description of the new license.

See http://llvm.org/docs/CodingStandards.html#file-headers
for a description of the new LLVM standard file header.

Note that since these files are not going into LLVM proper,
we are not adopting the changes to the file header, only the
license information.