Fix ICE in ArithmeticIF for INT*8 type
When half precision reals were introduced the table (aif) used in the
ILI codegen for ArithmeticIF was increased in size and the offset for
INT*8 increased by one. But the entry for half precision was not
inserted, this causes an ICE for INT*8 type. The fix is to introduce
the entry for the half precision type before the INT*8 type.