[DebugInfo] Fix for redundant lexical block in debug info
Flang was creating a redundant lexical block at the beginning of
of every subroutine. This was causing debuggers to incorrectly
assume the scope of a variable to be this lexical block.

This patch fixes this incorrect behavior by creating a lexical
block iff there is actually present in source code.