[DebugInfo] Corrected debuginfo for modules
Summary:
Flang was creating DIModule metadata without file and line number
information, as a result GDB is not able to find and print debug
info fortran modules.

This patch extends the flang to emit DIModule with  file and line
number information. This patch is dependent on LLVM 11.0 DIModule
metadata, since older versions of DIModule do not support file
and line information.