[runtime] Fix off-by-one error in minimum integers
runtime/flang/const.c defines min_int{1,2,4,8} to be -max_int{1,2,4,8}, which
is wrong for the two's complement representation. As a result, the MAXLOC
runtime function will compare each element in the input array with an incorrect
minimum value. If the array contains -huge(integer)-1, MAXLOC will return an
unexpected result.