Fix issue #670: Internal compiler error
The issue as reported is about the use of BOZ constants,
but the internal compiler error was generated by the handling
of the array constructor in the parameter definition.

This fix from the NVIDIA compiler addresses the issue.

Added two tests as well, from the original issue, and
a cut-down version used internally to identify the fix.