Fix private overwrite flag
find_def_in_most_recent_scope() set the private bit to 0 when it found an alias. This was most likely unintended and could cause issues when resolving multiple definitions of the same symbol at a later stage (if a private symbol appeared to be publicly available).

Tests for private flag overwrite added, based on flang github issue #890.