[flang1] Do not assume unempty derived types
When initializing an ACONST (a constant array or derived type instance), flang1
assumes that the ACL (array constructor list) has a non-NULL subict, which
normally represents the contents of the array or object. But a derived type may
be defined to contain no component part, which causes flang1 to crash when
dereferencing the NULL subict, unnecessarily.