Fix flang testing with newer gcc versions
When flang is being built with compiler version newer than
the one installed on the system, we need to make sure that
COMPILER_PATH variable is set for lit testing,
otherwise there might be compatibillity problems with the libgcc.a library.
Flang tests are building and executing programs, so some symbols
could be missing when those programs are linked with older libgcc.a
than the one used to build compiler.