[DebugInfo] Avoid data location of assumed shape array being optimized away
Base address of assumed shape array is marked as 'dataLocation' field of
array type, this is an artifical variable earlier mentioned as DW_TAG_variable.
With below options this artificial variable is optimized out.
%flang -g -S -emit-llvm test/debug_info/assumed_shape_noopt.f90
%llc -O0 -fast-isel=false -global-isel=false -filetype=obj assumed_shape_noopt.ll
Marking this variable as DW_TAG_formal_parameter is a right thing to do and also
avoids optimization.