Bug fix:Turn off constant propagation optimize of common block variable.
This patch fixes the error caused by constant propagation optimization on common block variables.

tools/flang1/flang1exe/optutil.c: If sym is a common block variable, do not optimize.
Instead it is handled in a way similar to how equivalence statements are handled.

tools/flang2/flang2exe/dtypeutl.cpp: Add a macro to check if a variable is a common block
variable and a function: returns true if two commom block variables are overlapping else false.

tools/shared/nmeutil.c: If sptr1 and sptr2 are overlapping common block variables,
do not optimize; instead it is handled in a way similar to how equivalence statements are handled.