[DebugInfo] line table entry missing for statements
flang was performing some opts even at "-g -O0" level.

Speifically, CSE is getting performed uncoditionally just before
the translation of the `store` instructions from ILI to LLVMIR.
Resulting in loss of debug-info and failures in GDB/gdb.fortran
testsuite. This patch inserts a check to disable the CSE opt if
it is not enabled.
Note: In flang CSE is enabled at `>=O1` optimization level.