[runtime] Fix OPEN and UNPACKC call delegation
Functions with 32-bit CLEN parameters, e.g. OPEN and UNPACKC, should
delegate to the respective versions with 64-bit CLEN parameters, i.e.
OPENA and UNPACKCA, instead of themselves.

Also fix a typo in the declaration of LDR_INTERN_INITA, ensuring that
the uppercase and lowercase versions of the symbol are consistent.