Do not escalate -Wsign-compare warnings to errors
Many -Wsign-compare warnings are caused by the fact that the STG_MEMBERS
macro defines a number of unsigned int fields, which are often used in
comparisons with signed variables. The problem is pervasive and requires
further investigation in the future. For now, we reduce the errors to
warnings by passing -Wno-error=sign-compare to Clang.

Also, make -DWITH_WERROR=on enable -Werror not only for C compilations
but also for C++ compilations.