Fix the bug that the name of ENTRY is used as an argument
Flang doesn’t catch and handle the syntax error that the
name of an entry point is used as an argument when a result
clause is present, ultimately causing a segmentation fault
in flang2. This patch fixes the issue by catching and reporting
the syntax error described above. It then restores the right state.

tools/flang1/flang1exe/semant3.c : If the name of an
entry point is used as an argument when there is a
result clause, a proper error message is reported and
the variable `sptr` is restored a correct value