Fix a segmentation fault related to assumed length character functions.
When an assumed length character function is not external,
its symbol type will at first be set to `ST_IDENT` and the
`AUTOBJ` field will be set to `1`. When a invocation of it
is encountered, its symbol type will be modified to `ST_PROC`.
But its `AUTOBJ` remains `1` which will causes a segmentation
fault later.

This patch fixes the issue by removing the function from the
automatic data list and setting its `AUTOBJ` to `0`.

The other issue is that flang sets the return length as `0`
for assumed length character functions, causing an ICE afterwards.

This patch fixes it by modifying the ast type of the
function (from ‘A_FUNC’ to `A_CALL`), and adding code to process
the argument for the length.