Prevent illegal argument accessing during procedure exporting.
When exporting a procedure, if `PARAMCT` is not positive (e.g.,
when a name of a function is used as the procedure interface in
a procedure declaration statement before the function is defined
and later the function declartion statement has a `RESULT` clause
and zero dummy arguments, `PARAMCT` will be 0), then it should be
illegal to access the beginning of the argument list of the function.
zero or negative  Failing to prevent this kind of illegal access
will cause a segmentation fault.

This patch fixes the bug by checking whether `PARAMCT`
is greater than 0 before accessing the argument list.