Annotate pure functions with the new 'willreturn' attribute
Recently a new attribute was introduced to LLVM. As it turned out,
we need to annotate the call sites for runtime library math functions
with this new attribute, otherwise we will miss crucial optimization
opportunities.

The 'pure' Fortran functions seem to be a reasonable approximation
of LLVM's 'willreturn' functions. Apparently, the most critical
math functions are 'pure' and therefore they are marked with the
'nounwind' and 'readnone' attributes.

This patch extends this marking with the new 'willreturn' attribute.

Signed-off-by: Paul Osmialowski <pawel.osmialowski@arm.com>