Fix the ICE caused by source-allocating a class(*) array from an array of deferred length character strings. See issue #713.
It also fixes the length parameter not being correctly
assigned problem mentioned at the end of issue #713.

This patch piggybacks the logic to handle
source-allocating a class(*) array from an array of
deferred length character strings on the code
handling source-allocating an unlimited_polymorphic
array from a string.

Use DDTG(dtype) instead of dtype to determine the source
type of a polymorphic allocation, in order to accommodate
a string array as well as a single string.

To prevent an internal compiler error, do not call
chkstruct if the element type of an array-type
source is deferred-length string,and do not set rank,
lsize, gsize, and kind for the allocatable destination
to avoid the length of the array being set as 0.