Fix 2 bugs related to passing a polymorphic array as an argument to a function
This patch fixes 2 errors:
1. When a polymorphic array is passed as an argument,
flang directly uses the element type in the
declaration statement of the dummy array to create
the function calling node.
2. When a section of a polymorphic array is passed
as an argument, flang doesn't get the first element
correctly.

This patch fixes the 1st bug by adding a new ast node
to obtain the actual element type if necessary and it
fixes the 2nd error by calling `get_poly_element_arg`
to get the correct first element.