Fix bugs pertaining to an array constructor with an implied-do containing an expression that is a function call returning an array.
When an implied-do loop in an array constructor
contains in its value list a function call that
returns an array, flang at present handles the
function call once outside the loop and then inserts
the returned values into the final array repeatedly
during the loop, causing wrong result.

This patch fixes the bug by invoking the function
and inserting the results in each iteration of
the loop.