[flang1] Fix bug in poly str pointer
The polymorphic variable is the member of one derived type and the pointer
pointing to string target or string pointer. The type descriptor (td) is
used to initialize the poly descriptor, but the section descriptor (sd)
is used to store the type info of string. Use td to store the type info
of string.