[flang1] Fix bug in searching RTE_template call
A RTE_set_intrin_type call is generated after generating RTE_template
in function emit_alnd when the variable is a derived type member.
The previous statement of "std" is the RTE_set_intrin_type call, which
is STD_PREV(std). So, the RTE_template call is STD_PREV(STD_PREV(std))
in this case.