[flang1] semutil.c::mkvarref(). Fix. (#1016)
Adding ST_MODPROC to switch() options in mkvarref() to fix an issue
relating Pure Function not being properly detected.

The issue occours because the pure function type has assignated
ST_MODPROC and is not recognized.
In the case of the Workaround, the type is ST_PROC. Debug reveald
that the assignated types are correct, but the probem is that
ST_MODPROC has no action defined, hence an error in the compiler.
The soluton is to add ST_MODPROC into the same section as ST_PROC
in order to perfom the same actions