[Flang1] Fix bug in exporting iso_fortran_env
The functions compiler_options and compiler_version are in
iso_fortran_env module. They are exported in iso_c_binding library
incorrectly when writing mod files. The stype of the two functions is
set to ST_INTRIN incorrectly when they are used again by importing the
incorrect library, which leads to an error of undeclared symbols.

Based on a patch originally by Duanyang Jing.