[flang2] Fix member_with_offset in outliner.cpp
Commit fdd777cee941eb7df368a54abad2de4dff744194 added some code that creates
a NME for an address in the uplevel table (conceptually a struct with many
anonymous fields) if one does not exist, in order to make the dumps easier
to read (presumably there would otherwise be many stores associated with
the same NME, which would be confusing). However, instead of looking for
a struct member with a matching offset, the function member_with_offset
clobbers the ADDRESS field for the first SPTR in the given linked list
of members. This patch corrects the function, and also fixes some formatting
problems in the code that calls member_with_offset.