[runtime] Use const char * where applicable; NFCI
This patch fixes many -Wincompatible-pointer-types-discards-qualifiers
and -Wcast-qual warnings that resulted from mixing char * variables with
const char * values, especially string literals. Most of these can be
trivially fixed by changing the types of the variables or functions in
question to const char *. Some cases are handled by calling strdup()
to put a literal in a non-const variable.

A couple of places in runtime/flang/eoshift.c cast string literals and
long long int * values into char *. They are left as FIXMEs for now as
the results of the casts are passed to many functions via char *
parameters, and fixing all of those functions will not be trivial.