[Debug] Display common block variable in sub-routines
Issue: Common block variables are shown as “optimized out” in subroutines. (Due to missing metadata references).
Fix: In the  “write_comm” function, instead of emitting the global variable with defn to IR, we capture the same into a member “cmblk_init_data” of variable agb.  At the end of processing a source file, in the function “assemble_end”, we will generate the same with all the debug info metadata references.