[flang2] Make flang aware of LLVM Target features
Add a -target_features option to flang2 that takes a comma-separated list of
LLVM target features suitable for embedding in LLVM metadata.

Teach flang2 to output this as a "target-features" metadata.

This is needed for flang to produce llvm objects that can be used with LTO and
a combination of targets, for example Arm SVE targets.

Signed-off-by: Richard Barton <richard.barton@arm.com>