Fix for the issue #872: Procedure pointers to internal procedures
This change is based on the solution discussed in #872.

The problem started to occur after introducing a very short commit
82683130aeebdbe6d5c70044cd79ef22e54f4f22. As a results, some of
the programs calling subroutines by pointers to internal procedures
started to segfault (examples are given in #872). The change
introduced by that commit was incorrect and had to be reverted
in favour of addressing the issue in other place.

Signed-off-by: Paul Osmialowski <pawel.osmialowski@arm.com>