[flang2] Fix bug in implied DO expression in a used module
The variables in a common block from a module will be copied into
current scope when a USE exists. However, the dinit feature of those
copied variables was removed when lowering symbols in flang1. This
patch suppresses the transformation of those copied variables to the
corresponding constant initialized in the common block, since we cannot
get their initialization info anyway.