cgmain: write target features and vscale range according to the compilation mode
The old code was emitting those function attributes always the same
way, no matter what compilation mode (debug or not) the compiler
is operating on.

This commit corrects this behavior. With debug mode, those
function attributes are added to the attributes block, along
with other attributes emitted by the write_function_attributes()
function.

Signed-off-by: Paul Osmialowski <pawel.osmialowski@arm.com>