[QuadFP][runtime] Make __BIGREAL_T 128-bit on supporting targets
__BIGREAL_T is intended to be the largest supported floating-point type on
a given target. Previously it was universally equivalent to __REAL8_T, but
now on targets that support quad-precision FP, we can typedef it to
__REAL16_T, or equivalently, float128_t. Doing so allows us to keep using
__BIGREAL_T where it makes sense, and reduces divergence in the runtime code.

float128_t is equivalent to __float128 on targets that provide that type
(PowerPC and X86), or long double on AArch64 and other targets that do not
provide __float128.