[QuadFP][libpgmath] Use float128_t instead of long double
Previously, quad-precision FP intrinsics were added to libpgmath
which used the long double type for parameters and return values,
assuming that the type was 128-bit. This patch introduces a
float128_t which is a typedef for __float128 on targets that provide
that type, or long double or targets that do not, including AArch64
(on which the long double type is in fact 128-bit).