Fix for regression introduced by PR #1101.
The root cause is that when an empty derived type parameter is declared in module,
an ACL structure for that parameter is generated, but its subc node is empty and
exported to the MOD file. But when importing the MOD file, flang cannot get the
information of this empty subc node.

This commit adds a flag to the MOD file to indicate whether the ACL structure
for the derived type parameter has a subc node.