[NFC] Add -fopenmp to test cases which use openmp runtime routines
Previously, classic flang provides ompstub for cases using openmp
runtime routines, but not adding the option "-fopenmp". gfortran/ifort
report errors of "undefined reference to ...".

This adds the option "-fopenmp" for the two test cases, which is
prepared for removing the "-lompstub" in the driver. After that, the two
test cases compiled without the option "-fopenmp" will report the error.