[Runtime] Move init_nomp.c to flangrti
The functions in init_nomp.c was added in libompstub.so. The ompstub
linking "-lompstub" is planned to be removed in the driver. However, it
is unknown how the functions in init_nomp.c are used. For example, as
the description in ilmtp.n, "_mp_lcpu2" is used for sections directive,
but there is no such runtime function generated for a OpenMP sections
program. To be safe, move init_nomp.c to flangrti.