[runtime] Avoid -Wdeprecated-non-prototype warnings; NFC
Delete extern function declarations that already have conflicting prototypes
declared earlier/elsewhere. This avoids errors such as:

.../flang/tools/flang1/flang1exe/induc.c:517:18: error: a function declaration
without a prototype is deprecated in all versions of C and is treated as a
zero-parameter prototype in C2x, conflicting with a previous declaration
[-Werror,-Wdeprecated-non-prototype]
  extern LOGICAL bv_mem();
                 ^
.../flang/tools/flang1/flang1exe/optimize.h:749:9: note: conflicting prototype
is here
LOGICAL bv_mem(int *, int);

In some cases, the function declaration is modified to complete the prototype.