[runtime] Replace or remove uses of undefined LINUX8664 macro
Various files under runtime/flang use the macro LINUX8664 to guard X86-only
code; but this is incorrect and ineffective because LINUX8664 is never defined
by any CMakeLists.txt (it is only defined when building runtime/libpgmath).
Necessary uses of the macro are modified to use TARGET_X8664 and/or
TARGET_LINUX_X8664, and unnecessary uses that never had any effect are simply
deleted.

This patch also guards some Windows-specific code with the WIN64 macro to avoid
compilation warnings.