[runtime] Define overloaded function types (comm.c, olap.c, scatter.c); NFCI
Classic Flang code uses tables of functions (with different argument types)
to achieve polymorphism. Previously, such a function was declared as
`void (*p)()`, to allow any function to be assigned to the function pointer
variable `p`. But when a call to such a function pointer is attempted,
Clang 15 now emits warnings like these:

.../flang/runtime/flang/comm.c:97:14: error: passing arguments to a function
without a prototype is deprecated in all versions of C and is not supported
in C2x [-Werror,-Wdeprecated-non-prototype]
    sk->start(sk->arg, rb, sb, rd, sd);
             ^
.../flang/runtime/flang/comm.c:117:13: error: passing arguments to a function
without a prototype is deprecated in all versions of C and is not supported
in C2x [-Werror,-Wdeprecated-non-prototype]
    sk->free(sk->arg);
            ^

To work around these issues, we replace polymorphic pointer parameters in such
functions with `void *` parameters, and cast the void pointers in the function
body to the intended type of the argument. By creating `typedef`s of the fully
prototyped function pointers, some degree of type checking is possible without
too much overhead. Relaxing the parameter type to `void *` also eliminates
`-Wincompatible-function-pointer-types` warnings when assigning functions
to function pointers.

This is the first patch in a series. The function pointers `start` and `free`
in the communication schedule type `sked` are updated.