[runtime] Specify location of libomp dependency
This patch makes it possible to run flang without setting
LD_LIBRARY_PATH for the libomp dependency in libflang and
libflangrti. This patch adds an $ORIGIN rpath to libflang and
libflangrti so that it will search for libomp in the same
directory as where these libraries are located. Typically, the
libraries for a toolchain are installed in the same directory.