[workflow] Fix push workflow in build_flang_arm64.yml
build_flang_arm64.yml has custom code for cloning the repository, since the workflow
runs in a container on a hosted runner. The workflow needs to be given an explicit
branch name to check out, otherwise it would always checkout the master branch.
This patch adds back the missing branch name to fix a push workflow issue on the
legacy branch, and also shuffles the steps slightly (NFC) to make build_flang.yml
and build_flang_arm64.yml more similar.