win: Remove '-mlong-double-128' compiler option
On Windows 'long double' type size is 64 bits, same as double.
As a workaround we started to use clang's '-mlong-double-128' option
to force 'long double' to be 128 bits.
With newer version of Windows SDK it causes an assertion in one of
system header.
Currently it breaks the Windows build, so it needs to remove.

Issue: #1334