[flang1] Fix a regression about array constructor
When the array constructor element is array of dynamic char, the
length can only be checked in runtime, in which case we cannot use
`eq_dtype` to check if the two dtypes are equal since the length could
be two different asts containing the length info.

This bug should exist long time ago, and it was hidden by 0918a2f2d.
It was exposed again in a545a61c9.