Fix issue 1336. incorrect results with entry statements.
When BIND(C) attribute is specified, store_return_value_for_entry function is returning
the pointer instead of scalar return. This resulted in invalid return value.

An explicit check is added to check the type of result operand and return the stored
value if the operand type is of pointer type.

The complex type is handled as double in x86 ABI. This is also fixed to generate
complex type instead of double. There may be some changes required in AARCH64 and
powerpc ABIs as well to handle the complex type in similar way. It is not clear
to me if at all any changes are required because the flow in AARCH64 and POWERPC
are different than X86.