[Flang1] Fix the string operation of file name
This bug came from 0bcbadbf3f7f056b298a7724b9689ebacbdae4e9, which uses
strdup to allocate non-const versions of const string, but does not
change the use of the original string. The use of original string (file
name) terminates the string by removing the last directory. So, we need
to replace the use of original string with the allocated string.