[flang2] byval and sret attributes should have associated types
When the return value of the function is a structure, the ABI might require
that the caller pass the structure into the function call as an implicit
input argument; in such cases the generated IR for the argument will
have the sret attribute. Similarly, pass-by-value arguments are given
the byval attribute in the IR. Since the adoption of opaque pointers,
such attributes are required to be accompanied with the pointee type. This
patch fixes flang2 to generate correct IR for such argument/return types.

Co-authored-by: Liuyunlong <liuyunlong16@huawei.com>